// web/src/shared/api/generated/model/paymentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  Pending: 'Pending',
  Succeeded: 'Succeeded',
  Failed: 'Failed',
  Refunded: 'Refunded',
} as const;